% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elixir4r-package.R
\docType{package}
\name{elixir4r-package}
\alias{elixir4r}
\alias{elixir4r-package}
\title{elixir4r: Run Elixir in R Markdown}
\description{
elixir engine for knitr
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/qiushiyan/elixir4r}
  \item Report bugs at \url{https://github.com/qiushiyan/elixir4r/issues}
}

}
\author{
\strong{Maintainer}: Qiushi Yan \email{qiushi.yann@gmail.com}

}
\keyword{internal}
