% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elixir.R
\name{ex_eval}
\alias{ex_eval}
\title{elixir interfaces}
\usage{
ex_eval(code)
}
\arguments{
\item{code}{a single elixir expression}
}
\description{
\code{ex_eval} for run string commands, \code{ex_eval} run
Evaluate string commands
}
