% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elixir.R
\name{ex_run}
\alias{ex_run}
\title{Run elixir file}
\usage{
ex_run(file, append_print = FALSE)
}
\arguments{
\item{file}{file path}

\item{append_print}{boolean, if IO.inpsect last line}
}
\description{
Run elixir file
}
